+++ packages/lib/src/index.ts
@@ -1,3 +1,3 @@
 export { transformerTypescriptDsl } from "./transform";
-export * from "./workaround";
+export * from "./extra/dummy";
 export * from "./extra/preset-icons";
